backend_api/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
yfinance
requests
psycopg2-binary
sqlalchemy
